Split crate in two
