--- README.tpl
@@ -1,7 +1,0 @@
-# [{{crate}}](https://crates.io/crates/juniper-from-schema)
-
-{{readme}}
-
----
-
-License: {{license}}
