+++ juniper-from-schema-code-gen/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+version = "0.2.0"
+authors = ["David Pedersen <david.pdrsn@gmail.com>"]
+categories = ["web-programming"]
+description = "Internal code generation crate for juniper-from-schema"
+documentation = "https://docs.rs/juniper-from-schema-code-gen"
+edition = "2018"
+homepage = "https://github.com/davidpdrsn/juniper-from-schema"
+keywords = ["web", "graphql", "juniper"]
+license = "MIT"
+name = "juniper-from-schema-code-gen"
+readme = "README.md"
+repository = "https://github.com/davidpdrsn/juniper-from-schema.git"
+
+[dependencies]
+syn = { version = "0.15.22", features = ["full"] }
+quote = "0.6.10"
+graphql-parser = "0.2.2"
+proc-macro2 = "0.4.24"
+heck = "0.3.0"
+regex = "1.1.0"
+lazy_static = "1.1.0"
+rustfmt-nightly = { version = "1.0.1", optional = true }
+colored = "1.8.0"
+
+[features]
+default = []
+format-debug-output = ["rustfmt-nightly"]
+
+[dev_dependencies]
+version-sync = "0.8"
+juniper = "0.11.0"
+
+[lib]
+proc-macro = true
+path = "src/lib.rs"
