+++ juniper-from-schema-code-gen/README.md
@@ -1,0 +1,5 @@
+# juniper-from-schema-code-gen
+
+Internal crate for [juniper-from-schema](https://crates.io/crates/juniper-from-schema).
+
+You shouldn't have to depend on this crate directly. The procedural macros are re-exported by [juniper-from-schema](https://crates.io/crates/juniper-from-schema).
