+++ juniper-from-schema-code-gen/src/ast_pass/ast_data_pass/find_special_scalar_types.rs
