+++ juniper-from-schema-code-gen/src/ast_pass/code_gen_pass.rs
@@ -202,7 +202,9 @@ impl<'doc> CodeGenPass<'doc> {
                 }
                 TypeKind::Type => {
                     let query_trail_type = ident(&field.inner_type);
-                    let trail = quote! { &QueryTrail<'a, #query_trail_type, Walked> };
+                    let trail = quote! {
+                        &QueryTrail<'a, #query_trail_type, juniper_from_schema::Walked>
+                    };
                     quote! {
                         /// Field method generated by `juniper-from-schema`.
                         fn #field_name<'a>(
