+++ juniper-from-schema-code-gen/src/ast_pass/code_gen_pass/gen_query_trails.rs
@@ -38,6 +38,8 @@ impl<'doc> CodeGenPass<'doc> {
 
     fn gen_query_trail(&mut self) {
         self.extend(quote! {
+            pub use juniper_from_schema::{Walked, NotWalked};
+
             /// A wrapper around a `juniper::LookAheadSelection` with methods for each possible child.
             ///
             /// Generated by `juniper-from-schema`.
@@ -47,40 +49,32 @@ impl<'doc> CodeGenPass<'doc> {
                 walked: K,
             }
 
-            impl<'a, T> QueryTrail<'a, T, NotWalked> {
+            impl<'a, T> QueryTrail<'a, T, juniper_from_schema::NotWalked> {
                 /// Check if the trail is present in the query being executed
                 ///
                 /// Generated by `juniper-from-schema`.
-                pub fn walk(self) -> Option<QueryTrail<'a, T, Walked>> {
+                pub fn walk(self) -> Option<QueryTrail<'a, T, juniper_from_schema::Walked>> {
                     match self.look_ahead {
-                        Some(inner) => {
-                            Some(QueryTrail {
-                                look_ahead: Some(inner),
-                                node_type: self.node_type,
-                                walked: Walked,
-                            })
-                        },
+                        Some(inner) => Some(QueryTrail {
+                            look_ahead: Some(inner),
+                            node_type: self.node_type,
+                            walked: juniper_from_schema::Walked,
+                        }),
                         None => None,
                     }
                 }
             }
 
-            /// A type used to parameterize `QueryTrail` to know that `walk` has been called.
-            pub struct Walked;
-
-            /// A type used to parameterize `QueryTrail` to know that `walk` has *not* been called.
-            pub struct NotWalked;
-
             trait MakeQueryTrail<'a> {
-                fn make_query_trail<T>(&'a self) -> QueryTrail<'a, T, Walked>;
+                fn make_query_trail<T>(&'a self) -> QueryTrail<'a, T, juniper_from_schema::Walked>;
             }
 
             impl<'a> MakeQueryTrail<'a> for juniper::LookAheadSelection<'a, juniper::DefaultScalarValue> {
-                fn make_query_trail<T>(&'a self) -> QueryTrail<'a, T, Walked> {
+                fn make_query_trail<T>(&'a self) -> QueryTrail<'a, T, juniper_from_schema::Walked> {
                     QueryTrail {
                         look_ahead: Some(self),
                         node_type: std::marker::PhantomData,
-                        walked: Walked,
+                        walked: juniper_from_schema::Walked,
                     }
                 }
             }
@@ -167,7 +161,7 @@ impl<'doc> CodeGenPass<'doc> {
                     /// Walk the trail into a field.
                     ///
                     /// Generated by `juniper-from-schema`.
-                    pub fn #name(&self) -> QueryTrail<'a, #field_type, NotWalked> {
+                    pub fn #name(&self) -> QueryTrail<'a, #field_type, juniper_from_schema::NotWalked> {
                         use juniper::LookAheadMethods;
 
                         let child = self.look_ahead.and_then(|la| la.select_child(#string_name));
@@ -175,7 +169,7 @@ impl<'doc> CodeGenPass<'doc> {
                         QueryTrail {
                             look_ahead: child,
                             node_type: std::marker::PhantomData,
-                            walked: NotWalked,
+                            walked: juniper_from_schema::NotWalked,
                         }
                     }
                 }
