+++ juniper-from-schema/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+version = "0.2.0"
+authors = ["David Pedersen <david.pdrsn@gmail.com>"]
+categories = ["web-programming"]
+description = "Generate Juniper code from you GraphQL schema"
+documentation = "https://docs.rs/juniper-from-schema"
+edition = "2018"
+homepage = "https://github.com/davidpdrsn/juniper-from-schema"
+keywords = ["web", "graphql", "juniper"]
+license = "MIT"
+name = "juniper-from-schema"
+readme = "README.md"
+repository = "https://github.com/davidpdrsn/juniper-from-schema.git"
+
+[dependencies]
+juniper-from-schema-code-gen = { version = "0.2.0", path = "../juniper-from-schema-code-gen" }
+juniper = "0.11.0"
+
+[dev_dependencies]
+serde_json = "1.0.33"
+assert-json-diff = "0.2.1"
+maplit = "1.0.1"
+version-sync = "0.8"
+trybuild = "1.0.3"
