+++ juniper-from-schema/examples/enumeration_types.rs
