+++ juniper-from-schema/examples/interface.rs
