+++ juniper-from-schema/examples/query_trails.rs
