+++ juniper-from-schema/examples/union_types.rs
