+++ juniper-from-schema/tests/compile_fail/enum.stderr
