+++ juniper-from-schema/tests/compile_pass/empty_mutations.rs
