+++ juniper-from-schema/tests/compile_pass/enums.rs
