+++ juniper-from-schema/tests/compile_pass/input_object.rs
