+++ juniper-from-schema/tests/compile_pass/ownership_attributes.rs
@@ -1,7 +1,9 @@
 #![allow(dead_code, unused_variables, unused_must_use, unused_imports)]
 include!("../setup.rs");
 
-graphql_schema_from_file!("../../../tests/schemas/ownership_attributes.graphql");
+graphql_schema_from_file!(
+    "../../../juniper-from-schema/tests/schemas/ownership_attributes.graphql"
+);
 
 pub struct Query;
