+++ juniper-from-schema/tests/compile_pass/query_trail.rs
