+++ juniper-from-schema/tests/compile_pass/query_trail_methods_for_union_types.rs
