+++ juniper-from-schema/tests/compile_pass/simple_non_null_scalars.rs
