+++ juniper-from-schema/tests/compile_pass/simple_nullable_scalars.rs
