+++ juniper-from-schema/tests/schemas/complex_schema.graphql
