+++ juniper-from-schema/tests/schemas/doc_schema.graphql
