+++ juniper-from-schema/tests/schemas/original_complex.graphql
