+++ juniper-from-schema/tests/schemas/ownership_attributes.graphql
