+++ juniper-from-schema/tests/schemas/returning_references.graphql
