+++ juniper-from-schema/tests/schemas/very_simple_schema.graphql
